Removes block syntax insertion on the transclusion
When writing the documentation, it was harder to see where the inserted
json would appear, as the ```json block quotes would only show after
building.

This commit removes the quoting from the insertion, and moves this
responsability to the markdown so we can see where things will fit when
writing the documentation.