Expands the introductory paragraph on Telemetry API
This commit adds more context around the Telemetry API service. It
informs that it is backed by
[elastic-harvestjs](https://github.com/agco/elastic-harvesterjs), which
lets you manipulate the information sent back with queries to return
precise information.

It also indicates the documentation on [ACGO's JSON API
Profiles](https://github.com/agco/agco-json-api-profiles) to the reader,
so they have more information on the extra possibilities to manipulate
the returned information.