Extends the lexer to style custom tags
We would like to have a curl example and the schema for the response on
the Equipment API.

The schema itself tend to be large and add a bit of visual noise on the
page. We would like to let the user only see the schema when it desires,
hiding with the `language_tabs` feature of Slate.

This commit initiates the setup, extending the higlighting for our
custom language tags, so schema has the same highlight as JSON, and curl
extends shell.