Introduces custom tags on Telemetyr API content
We would like to have a curl example and the schema for the response on
the Equipment API.

The schema itself tend to be large and add a bit of visual noise on the
page. We would like to let the user only see the schema when it desires,
hiding with the `language_tabs` feature of Slate.

This commit initiates converts the blockquotes to HTML elements so we
can use the classes to hide according to the current language, adds curl
examples on the Telemetry API, and converts the coding quotes into the
custom tags previously added.